Refactor view model to view mapping and update EnumProvider usage in XAML files
