+++ samples/TestApp/TestApp/Samples/SlimDataGrid/SlimDataGridView.axaml
@@ -2,10 +2,10 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:sdg="clr-namespace:Zafiro.Avalonia.Controls.SlimDataGrid;assembly=Zafiro.Avalonia"
              xmlns:avalonia="https://github.com/projektanker/icons.avalonia"
              xmlns:slimDataGrid="clr-namespace:TestApp.Samples.SlimDataGrid"
              xmlns:c="clr-namespace:Zafiro.Avalonia.Controls;assembly=Zafiro.Avalonia"
+             xmlns:misc="clr-namespace:Zafiro.Avalonia.Misc;assembly=Zafiro.Avalonia"
              mc:Ignorable="d" d:DesignWidth="400" d:DesignHeight="450"
              x:Class="TestApp.Samples.SlimDataGrid.SlimDataGridView"
              x:DataType="slimDataGrid:SlimDataGridViewModel">
@@ -18,14 +18,10 @@
         <StyleInclude Source="/Samples/SampleStyles.axaml" />
 
         <Styles>
-            <!-- Elimina subrayado en todas las HyperlinkButton -->
             <Style Selector="HyperlinkButton">
                 <Setter Property="TextBlock.TextDecorations" Value="{x:Null}" />
                 <Setter Property="CornerRadius" Value="0" />
                 <Setter Property="BorderThickness" Value="0" />
-                <!-- <Setter Property="PointerOverTextDecorations" Value="None"/> -->
-                <!-- <Setter Property="PointerPressedTextDecorations" Value="None"/> -->
-                <!-- <Setter Property="FocusTextDecorations" Value="None"/> -->
             </Style>
         </Styles>
     </UserControl.Styles>
@@ -37,7 +33,7 @@
         <DataTemplate x:Key="SurnameTemplate">
             <TextBlock TextAlignment="Center" Foreground="#536A72" x:DataType="x:String" Text="{Binding }" VerticalAlignment="Center" />
         </DataTemplate>
-        <ControlTheme x:Key="RowTheme" TargetType="sdg:Row" BasedOn="{StaticResource {x:Type sdg:Row}}">
+        <ControlTheme x:Key="RowTheme" TargetType="Row" BasedOn="{StaticResource {x:Type Row}}">
             <Setter Property="BorderThickness" Value="0 0 0 1" />
             <Setter Property="BorderBrush" Value="#4F000000" />
             <Setter Property="Padding" Value="10 5" />
@@ -68,43 +64,43 @@
                 </InlineUIContainer>
             </TextBlock>
 
-            <sdg:SlimDataGrid HorizontalAlignment="Center" Background="#FFC9EC" Padding="10"
-                              VerticalAlignment="Center"
-                              MinWidth="320"
-                              HeaderBackground="#20000000" BorderThickness="2"
-                              BorderBrush="Black" CornerRadius="10"
-                              ItemsSource="{Binding People}" HeaderPadding="10 10"
-                              RowTheme="{StaticResource RowTheme}">
-                <sdg:SlimDataGrid.Columns>
-                    <sdg:Column Header="Name" Binding="{Binding Name}"
-                                Width="*"
-                                CellTemplate="{StaticResource NameTemplate}" />
-                    <sdg:Column Header="Surname"
-                                Width="*"
-                                Binding="{Binding Surname}"
-                                CellTemplate="{StaticResource SurnameTemplate}">
-                        <sdg:Column.HeaderTemplate>
+            <SlimDataGrid HorizontalAlignment="Center" Background="#FFC9EC" Padding="10"
+                          VerticalAlignment="Center"
+                          MinWidth="320"
+                          HeaderBackground="#20000000" BorderThickness="2"
+                          BorderBrush="Black" CornerRadius="10"
+                          ItemsSource="{Binding People}" HeaderPadding="10 10"
+                          RowTheme="{StaticResource RowTheme}">
+                <SlimDataGrid.Columns>
+                    <Column Header="Name" Binding="{Binding Name}"
+                            Width="*"
+                            CellTemplate="{StaticResource NameTemplate}" />
+                    <Column Header="Surname"
+                            Width="*"
+                            Binding="{Binding Surname}"
+                            CellTemplate="{StaticResource SurnameTemplate}">
+                        <Column.HeaderTemplate>
                             <DataTemplate DataType="x:String">
                                 <TextBlock HorizontalAlignment="Center" Text="{Binding}" />
                             </DataTemplate>
-                        </sdg:Column.HeaderTemplate>
-                    </sdg:Column>
-                    <sdg:Column Header="Link"
-                                Width="*">
-                        <sdg:Column.HeaderTemplate>
+                        </Column.HeaderTemplate>
+                    </Column>
+                    <Column Header="Link"
+                            Width="*">
+                        <Column.HeaderTemplate>
                             <DataTemplate DataType="x:String">
                                 <TextBlock HorizontalAlignment="Center" Text="{Binding}" />
                             </DataTemplate>
-                        </sdg:Column.HeaderTemplate>
-                        <sdg:Column.CellTemplate>
+                        </Column.HeaderTemplate>
+                        <Column.CellTemplate>
                             <DataTemplate>
                                 <Button HorizontalAlignment="Center" avalonia:Attached.Icon="fa-solid fa-arrow-up-right-from-square" CommandParameter="http://www.youtube.com"
-                                        Command="{Binding Source={x:Static Commands.Instance}, Path=LaunchUri}" />
+                                        Command="{Binding Source={x:Static misc:Commands.Instance}, Path=LaunchUri}" />
                             </DataTemplate>
-                        </sdg:Column.CellTemplate>
-                    </sdg:Column>
-                </sdg:SlimDataGrid.Columns>
-            </sdg:SlimDataGrid>
+                        </Column.CellTemplate>
+                    </Column>
+                </SlimDataGrid.Columns>
+            </SlimDataGrid>
 
         </DockPanel>
     </c:HeaderedContainer>
