+++ src/Zafiro.Avalonia.DataViz/Heatmaps/HeatmapConverter.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Avalonia;
 using Avalonia.Data.Converters;
 using Avalonia.Media;
-using Zafiro.Avalonia.ViewLocators;
+using Zafiro.Avalonia.Misc;
 
 namespace Zafiro.Avalonia.DataViz.Heatmaps;
