+++ src/Zafiro.Avalonia/Behaviors/PointerPressedOutsideTriggerBehavior.cs
@@ -2,7 +2,7 @@ using System.Reactive.Disposables;
 using Avalonia.Input;
 using Avalonia.VisualTree;
 using Avalonia.Xaml.Interactivity;
-using Zafiro.Avalonia.ViewLocators;
+using Zafiro.Avalonia.Misc;
 
 namespace Zafiro.Avalonia.Behaviors
 {
