+++ src/Zafiro.Avalonia/Misc/BitmapFactory.cs
@@ -1,7 +1,7 @@
 ﻿using Avalonia.Media.Imaging;
 using Avalonia.Platform;
 
-namespace Zafiro.Avalonia.ViewLocators;
+namespace Zafiro.Avalonia.Misc;
 
 public static class BitmapFactory
 {
