+++ src/Zafiro.Avalonia/Misc/CollapseNameTrimming.cs
@@ -1,7 +1,7 @@
 ﻿using Avalonia.Media;
 using Avalonia.Media.TextFormatting;
 
-namespace Zafiro.Avalonia.ViewLocators;
+namespace Zafiro.Avalonia.Misc;
 
 public class CollapseNameTrimming : TextTrimming
 {
