+++ src/Zafiro.Avalonia/Misc/ColorInterpolator.cs
@@ -1,6 +1,6 @@
 ﻿using Avalonia.Media;
 
-namespace Zafiro.Avalonia.ViewLocators;
+namespace Zafiro.Avalonia.Misc;
 
 public static class ColorInterpolator
 {
