+++ src/Zafiro.Avalonia/Misc/Enum.cs
@@ -1,4 +1,4 @@
-﻿namespace Zafiro.Avalonia.ViewLocators;
+﻿namespace Zafiro.Avalonia.Misc;
 
 public class Enum
 {
