+++ src/Zafiro.Avalonia/Misc/EnumExtensions.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel;
 using System.Reflection;
 
-namespace Zafiro.Avalonia.ViewLocators;
+namespace Zafiro.Avalonia.Misc;
 
 // TODO: Move to Zafiro
 public static class EnumMixin
