+++ src/Zafiro.Avalonia/Misc/EnumItem.cs
@@ -1,3 +1,3 @@
-namespace Zafiro.Avalonia.ViewLocators;
+namespace Zafiro.Avalonia.Misc;
 
 public record EnumItem(System.Enum Value, string Description);
