+++ src/Zafiro.Avalonia/Misc/Link.cs
@@ -2,7 +2,7 @@ using System.Reactive;
 using CSharpFunctionalExtensions;
 using Zafiro.Avalonia.Services;
 
-namespace Zafiro.Avalonia.ViewLocators;
+namespace Zafiro.Avalonia.Misc;
 
 public class Link
 {
