+++ src/Zafiro.Avalonia/Misc/VisualExtensions.cs
@@ -1,6 +1,6 @@
 using Avalonia.VisualTree;
 
-namespace Zafiro.Avalonia.ViewLocators;
+namespace Zafiro.Avalonia.Misc;
 
 public static class VisualExtensions
 {
